# Separate fading props from pop-in props