### Related Issue
N/A

### Implementation
This is the consolidated code. The fully separated code had higher framerates, but also much higher variability. For the sake of performance stability, I chose to use this one. It also serves as a good middle ground. Benchmarking results at the bottom.

Add another vbsp compiler flag that separates props that fade from those that pop.
From there the fade code is modified to add an else if statement that checks if it has the pop flag, at which point it does less expensive calculations (ignores things it doesn't need like the distance before a prop starts fading).

Screenspace pop-in is a separate thing because i found that it was cheaper than merging it with screenspace fade, likely due to screenspace fade worrying about more variables. interestingly, this caused a performance increase even on maps _without_ screenspace fade enabled.

### Screenshots
N/A

### Checklist
<!-- You MUST answer "yes" to all of these to open a pull request -->
<!-- To tick a checkbox, place an 'x' in it, like so: [x] -->
- [x] No other PRs implement this idea.
- [x] This PR does not introduce any regressions.
- [x] I certify that this PR is my own entirely original work, or I have permission from and have attributed the relevant authors.
- [x] I have agreed to and signed this project's [Contributor License Agreement](https://cla-assistant.io/mastercomfig/team-comtress-2).
- [x] This PR only contains changes to the engine and/or core game framework
- [x] This PR targets the `community` branch.

<!-- You do NOT have to answer "yes" to the following, but please mark them if relevant -->
<!-- To tick a checkbox, place an 'x' in it, like so: [x] -->
- [ ] This change has been filed as an issue.
- [x] No other PRs address this.
- [x] This PR is as minimal as possible.
- [x] This PR does not introduce any regressions.
- [x] This PR has been built and locally tested on at least one platform.
- [ ] This PR has been tested on all platforms, on both a dedicated server and on a listen server.

### Testing Checklist
<!-- You do not have to test on all platforms to open a pull request -->
|         |            Client             |            Server             | Version                     |
|---------|:-----------------------------:|:-----------------------------:|-----------------------------|
| Windows | Built & Tested | N/A  | Windows 10 Home x64 Version 20H2 |
|   Linux | <!-- Built, Tested or N/A --> | <!-- Built, Tested or N/A --> | <!-- `uname -vr` output --> |
|  Mac OS | <!-- Built, Tested or N/A --> | <!-- Built, Tested or N/A --> | <!-- e.g. Catalina -->      |

### Caveats
In order to make room for the new flag, I had to comment out an older flag (`STATIC_PROP_WC_MASK`), which only had one reference, which was also commented out, so it ultimately went unused.

### Alternatives

- Completely separating the fading props from the pop-in ones. This caused higher performance but much higher variability.
- Consolidating the screenspace calculations for both. This resulted in lower framerates and higher variability.

### Performance

Averaged over 5 runs of a demo, full data can be found [here](https://github.com/mastercomfig/team-comtress-2/files/6536926/tc2.benchs.txt)
|         |             FPS             |         Variability         |        MS per frame         |            Time             |
|---------|:-----------------------------:|:-----------------------------:|:-----------------------------:|-----------------------------|
Current TC2 | 253.68 | 32.79 ms | 3.94 ms/f | 9.823 seconds |
Separate pop and fade | 261.84 | 34.52 ms | 3.81 ms/f | 9.517 seconds |
Consolidated pop and fade | 257.73 | 30.48 ms | 3.88 ms/f | 9.669 seconds |
Consolidated + screenspace | 256.50 | 31.26 ms | 3.89 ms/f | 9.715 seconds |

### Other notes

The specs of the computer that benchs were done on is:
- Win 10 Home x64 (Version 20H2)
- Intel I7-6700HQ
- Nvidia GTX 960M
- 16GB Ram

The bench i used for this was just me going around in spectator on a map i compiled (so i knew what kind of props were in it), but i'd really appreciate it if others would try benchmarking this PR vs current TC2 on some more active demos!